Implement an algorithm that generates rooms independently, then attempts to place them.
Inspired by this interview article:
https://www.rockpapershotgun.com/2015/07/28/how-do-roguelikes-generate-levels/

I haven't implemented everything in that article, and currently this
doesn't look very much better than what carve.pl can do, but I have
hopes for improving it.  If this really is what Brogue does, then it
has to be a viable approach, because Brogue's levels are frequently
quite beautiful (and I'm talking about the shape, here; the colors
are beautiful too, but that's not what I'm after at the moment).