Bugfix: diagonal lollipop stems should be gridbug-walkable.
This ensures the stems actually end up as part of the final level.
(Previously the fixup code was cleaning them up, leaving usually a
single extra floor tile sticking out of the side of a room as evidence
that the code had tried to put a corridor there.)

Also, insert || "" to suppress an annoying compare-to-undef warning.