Introduce a new mine generator, that works by actually mining minerals.
Seed the level with gold and gems embedded in solid rock, then send
several "miners" around the level to dig a bunch of them out.  Then do
the usual wall fixups and such, and add stairs.

I didn't bother adding floor items, monsters, or traps, but all of
that could work the same way as with existing level types.