Remove the big "X" artifact in the middle of the level.
The hardcoded initial goals created an artifact in the middle of each
level.  It frequently got obscured as the level developed, but when it
didn't, it was pretty glaring.  Randomize the initial goals (still
within each quadrant, but not always at the center of each quadrant),
so that there's no consistant X shape.

Also, parameterize the depth.